♻️ Move utils to utils file
Relates #45
